test setup was upgraded, a couple of Moose-related bugs came out
1.Had to remove use Typeconstraints from Varbind.pm. Field type was conflicting
with the Moose-provided method. No sideffects observed.

2.Had to remove the session field from SNMP::Class::Roles::Implementation
and duplicate it in NetSNMP.pm and Dummy.pm. Previously, the session was
being declared inside Implementation.pm and its type clarified by the former
two modules by using '+session'. This is not allowed for Moose Roles, so after
the Moose upgrade it was dying. So the session attribute is now declared on the
two 'specific' roles. Seems to be working.