-Fixed missing MIB enum mapping problem
In some cases, agents will return values for enumerated mib objects
which do not have a corresponding mapping. For example,
CISCO-VTP-MIB::managementDomainVersionInUse is defined as
INTEGER {version1(1), version2(2), none(3), version3(4)}
but some old agents will go ahead and return a 0 (which is not included
in the above mapping). This update fixes that problem by returning the
raw number. May visit this again. The proper thing to do would be to
emit an exception.