Facts and personalities system first commit
A system of roles representing different behaviors (e.g. 'a router' or
'a switch' has been implemented). Each role knows what oid are required,
has a predicate function to ascertain whether it is appropriate to apply
to an SNMP::Class object and a facts calculation routine where all the
logic eventually goes.