add ifIndex check for vtpVlanIfIndex
Some cisco switches report a vtpVlanIfIndex value that does not exist in
the interfaces table. This can happen when adding a layer-3 interface
and then delete it. The interface is created and a matching entry is
added in the vtp table, but it is not removed upon interface deletion.
Added a check in VtpVlanState personality to handle this case
gracefully.