Fix varbind-roles test
Previous test used the ifPhysAddress object to test whether the
S:C:Varbind::DisplayHint worked correctly. Unfortunately, the
mib parser in Net-SNMP no longer returns a hint for the
ifPhysAddress, possibly due to a change in a mib. An alternative
object, ipv6InterfaceIdentifier was chosen and a simple test was
added in place of the previous test. ipv6InterfaceIdentifier seems
to have a hint for the time being.