Refactor Serializable.pm to gracefully load Sereal
Use Module::Load, Module::Load::Conditional to test whether
Sereal is available and create appropriate serialize and
unserialize functions using Sub::Install. If Sereal is not
available, both functions will die when used.

An appropriate test for either case was added.

This framework will allow the addition of other serialization
modules which the code will be trying to load in a specific
order.