Fix bug in t/fact.t related to JSON order
Due to newer Perls no longer returning a hash in a deterministic
order (which is a documented feature), JSON probably alters the
order of the items inside serialized json objects. A test was
failing stochastically due to that fact, so a more robust test
handling all possible valid outcomes was created.