First version of generalized impl. plugins
A framework to formalize the creation of different implementations of
the same functionality is introduced.

The S:C:PluggableImplementations module can be just used from any parent
module wishing to have implementation plugins under its namespace. It is
assumed that the parent module is a Moose::Role and that all the
implementation submodules are using it as a base by invoking "with
parent_module"

Using the S:C:PluggableImplementations will scan under the including
module's namespace and load all the implementations found.

Then, it will try to sort the implementation submodules by the value of their
priority() subroutine and then try to find the first one having a true
value returned by their works() subroutine. Then, it will use
Sub::Install to install a plugin_meta subroutine to the parent module
which will be returning the meta object of the class of the selected
implementation submodule.

To use a parent module along with an automatically selected
implementation, one would have to just use parent_module and then do
something like parent_module->plugin_meta->apply( object ) to any object
wishing to acquire the specific moose role.