Work towards a Cacheable SNMP::Class
This is an intermediate commit. At present, a new approach to
serializing parts of an object consisting of multiple roles is
introduced. Each role has a serialize_rolename method and an
unserialize_rolename. The composing object is responsible for
using those methods to achieve the final expected behaviour.

The changes needed by SNMP::Class are not yet finished.

Significant changes were made in order for the Cache.pm
role to apply both the S:C:R:Cache and the S:C:R:Cache::Files.
Applying only the latter unfortunately does not introduce the
attributes of its ancestor to the targeted object. So to get all
the parent and child attributes one would have to first apply
the child role and then apply the parent role. This is all handled
by changes in the PluggableImplementations.pm module, which now
plants a new function named plugin_apply to any package that includes
it.