Add methods to dump varbind or SNMP::Class into hash
- Add to_hash method to dump a varbind into a hash.
- Add to_hashes method to dump all the varbinds inside a hash
into an array of varbind hashes. This method is not attached to
a resultset, because only a real session knows the host it is
talking to. The host of the session is attached as a key to each
individual varbind hash
- This hash is suitable to be fed into code that will convert it to
JSON and then e.g. insert into an ElasticSearch cluster, or Mongo,
or anything similar.