fix problem with Digest::SHA1 missing on Debian
FreeBSD always had the module Digest::SHA1 available as a port or
package. Unfortunately Debian (and Ubuntu I suspect) does not
have a Digest::SHA1 module. But it has the newer Digest::SHA,
which has exactly the same API. So, a new small module
SNMP::Class::SHA1 is introduced which tries to load the two
modules and uses whatever works first. It exports the sha1_hex
function and can is used as drop-in replacement for Digest::SHA1
'sha1_hex'.