better signal handling in Worker
Replaced the two signal handlers with a foor loop, better. Also,
inside a signal handler, when the signal is set to IGNORE, the
handler variable was declared local, so after the handler did
die or exit, the handler was reset to its original. That resulted
in signals caught two times, e.g. when the user typed Ctrl-C,
there was a SIGINT from the shell (all members in the process
group get it) and then there was a signal from the master itself
to the worker. So that produced some nasty warnings while dying.
Anyway, fixed by removing the local qualifier. Now IGNORE is there
to stay (albeit not for long since the worker is dying).