handle rare case of noSuchName from agent when getnext
In cases where we have issued a getnext to an agent and there exists
no next lexicographical object, the agent (per RFC1157) can return
error number 2, which means noSuchName. Although this library was
used for many years, never had this rare condition been encountered.
Trying to ask an old switch for laLoad triggered this condition. Our
code actually used to check the error number of getnext, but always
died when it was set.

We have used Throwable::Error and Try::Tiny to gracefully handle this
kind of corner case. The getnext throws an appropriate exception object
and the snmpwalk can examine it to determine whether we should actually
die of just print a debug message and finish the snmpwalk loop.